[OpenMP] Compute a proper module slice for the CGSCCC pass
Summary:
The module slice describes which functions we can analyze and transform
while working on an SCC as part of the CGSCC OpenMPOpt pass. So far, we
simply restricted it to the SCC. In a follow up we will need to have a
bigger scope which is why this patch introduces a proper identification
of the module slice. In short, everything that has a transitive
reference to a function in the SCC or is transitively referenced by one
is fair game.

Reviewers: jhuber6, fghanim, JonChesterfield, grokos, AndreyChurbanov, ye-luo, tianshilei1992, ggeorgakoudis

Subscribers: yaxunl, hiraditya, bollu, guansong, sstefan1, llvm-commits, openmp-commits

Tags: #llvm

Differential Revision: https://reviews.llvm.org/D83270