[OpenMP] Replace function pointer uses in GPU state machine
Summary:
In non-SPMD mode we create a state machine like code to identify the
parallel region the GPU worker threads should execute next. The
identification uses the parallel region function pointer as that allows
it to work even if the kernel (=target region) and the parallel region
are in separate TUs. However, taking the address of a function comes
with various downsides. With this patch we will identify the most common
situation and replace the function pointer use with a dummy global
symbol (for identification purposes only). That means, if the parallel
region is only called from a single target region (or kernel), we do not
use the function pointer of the parallel region to identify it but a new
global symbol.

Reviewers: jhuber6, fghanim, JonChesterfield, grokos, AndreyChurbanov, ye-luo, tianshilei1992, ggeorgakoudis

Subscribers: yaxunl, hiraditya, bollu, guansong, sstefan1, aaron.ballman, llvm-commits, openmp-commits

Tags: #llvm

Differential Revision: https://reviews.llvm.org/D83271